test location handler

+++ driver-location/server/types.go
@@ -2,12 +2,12 @@ package server
 
 type Location struct {
 	ID   string  `json:"id"`
-	Lat  float32 `json:"latitude"`
-	Long float32 `json:"longitude"`
+	Lat  float64 `json:"latitude"`
+	Long float64 `json:"longitude"`
 }
 
 type LocationUpdate struct {
 	UpdatedAt string  `json:"updated_at"`
-	Lat       float32 `json:"latitude"`
-	Long      float32 `json:"longitude"`
+	Lat       float64 `json:"latitude"`
+	Long      float64 `json:"longitude"`
 }
